requirements.txt
flask
flask_sqlalchemy
flask_socketio
eventlet
scikit-learn
pandas
numpy